Add cast for 64bit compilation
git-svn-id: https://svn.osgeo.org/gdal/trunk@15416 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/gdaldriver.cpp
@@ -1257,7 +1257,7 @@ int CPL_STDCALL GDALValidateCreationOptions( GDALDriverH hDriver,
                     {
                         CPLError(CE_Warning, CPLE_NotSupported,
                              "'%s' is of size %d, whereas maximum size for %s creation option is %d.",
-                             pszValue, strlen(pszValue), pszKey, atoi(pszMaxSize));
+                             pszValue, (int)strlen(pszValue), pszKey, atoi(pszMaxSize));
                         bRet = FALSE;
                     }
                 }
